requirements.txt
streamlit
seabord
numpy
pandas
matplotlib.pyplot